Move OVTextualInversionLoaderMixin in loaders.py
